moved class definitions from pundit_spec.rb to spec_helper.rb
